requirements.txt
boto3>=1.34.0
python-dotenv>=1.0.0
opensearch-py>=2.4.0
requests-aws4auth>=1.2.0
aws-sam-cli>=1.100.0
